kk_lab9/requirements.txt
selenium
pytest
pyyaml